fix parsing of yarn output
It appears that yarn (or the underlying node process abstractions)
do not guarantee that each line emitted by the process will be a
complete json blob.

This change buffers the output of the audit command into a single
string, then splits it on newlines and parses each line. This
ensures that no matter what chunks the yarn child process emits,
we'll parse complete JSON blobs